moved all paths to subfolders in the main directory to a single reactiveValues
Before the paths to subfolders within the main directory were hardcoded into the script where ever the subfolders were accessed. The paths are still hardcoded, but now they are at least only hardcoded in a single place, in the Initial Setup section. If any of these paths need to change in the future, it should be easier to change them.