display notification when saving file(s)
Did. I added notifications to the observeEvents for the save_survey and save_docs buttons. The message depends on whether one or both of the documents already exist in the file folder.

Didn't do. The other messages in the app are displayed at the top of the main panel with renderText. I did not replace any of these messages outside of the observeEvents for the save_survey and save_docs buttons.